feat(weaver_emit): add initial implementation of the emit library for OTLP signals
