+++ crates/weaver_emit/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "weaver_emit"
+version.workspace = true
+authors.workspace = true
+repository.workspace = true
+license.workspace = true
+publish.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+weaver_semconv = { path = "../weaver_semconv" }
+weaver_resolved_schema = { path = "../weaver_resolved_schema" }
+weaver_forge = { path = "../weaver_forge" }
+opentelemetry.workspace = true
+opentelemetry_sdk = { version = "0.27.1", features = ["rt-tokio"] }
+opentelemetry-otlp = "0.27.0"
+opentelemetry-stdout = { version = "0.27.0", features = ["trace"] }
+tokio = { version = "1.43.0", features = ["full"] }
+ordered-float.workspace = true
+
+[dev-dependencies]
+futures-util = { version = "0.3", default-features = false }
+
+[lints]
+workspace = true
+
+[package.metadata.cargo-machete]
+# force cargo machete to ignore the following crates
+# remove this section once this crate is integrated into the rest
+# of the project
+ignored = []
