+++ crates/weaver_emit/README.md
@@ -1,0 +1,5 @@
+# Weaver emit library
+
+Status: **Work-In-Progress**
+
+This crate provides the "emit" library for emitting OTLP signals generated from registries.
